ipa v0.0.1 IPA.Address

Functions for defining an IP address struct from a valid IP address.

The IP address struct contains the original dot-decimal notation address, it’s IP version, and the address as a binary value, a hexadecimal value, a binary dot notation, and a 4 element tuple.

Types

addr_struct :: %IPA.Address{address: String.t, bin: String.t, bits: String.t, block: atom, hex: String.t, octets: tuple, version: non_neg_integer}

The IP Address Struct

Functions

address(addr)

Specs

address(String.t) :: addr_struct | :invalid

Defines an IP Address struct.

Example

iex> {:ok, ip} = IPA.address("192.168.0.1")
{:ok,
%IPA.Address{address: "192.168.0.1", address_type: nil,
bin: "0b11000000101010000000000000000001",
bits: "11000000.10101000.00000000.00000001", block: :rfc1918,
hex: "0xC0A80001", octets: {192, 168, 0, 1}, version: 4}}
iex> ip.address
"192.168.0.1"
iex> ip.bin
"0b11000000101010000000000000000001"
iex> ip.bits
"11000000.10101000.00000000.00000001"
iex> ip.block
:rfc1918
iex> ip.hex
"0xC0A80001"
iex> ip.octets
{192, 168, 0, 1}
iex> ip.version
4
iex> IPA.address("192.168.0.256")
{:error, "Not a valid ip address"}
iex> IPA.address("192.168.0")
{:error, "Not a valid ip address"}

block(addr)

to_binary(addr)

to_bits(addr)

to_hex(addr)

to_octets(addr)

valid?(addr)