ipa v0.0.1 IPA.Address

Functions for working with IP addresses.

Currently only compatible with IPv4 addresses.

Types

addr :: String.t

addr_struct :: %IPA.Address{address: addr, binary: String.t, bits: String.t, block: atom, hex: String.t, octets: tuple, reserved: boolean, version: non_neg_integer}

Functions

address(addr)

Specs

address(addr) :: {:ok, addr_struct} | {:error, atom}

Defines an IP Address struct containing:

address - original dotted-decimal notation address

version - IP version

reserved - whether it is a reserved address or not

block - it’s address block, either public or one of a number of private blocks

binary - the address as a binary value

hex - the address as a hexadecimal value

bits - the address as binary dot notation

octets - the address as a 4 element tuple.

Returns {:ok, ip}, where ip is the struct defined above, or {:error, :invalid_ipv4_address} if the IP address is invalid.

Examples

iex> {:ok, ip} = IPA.address("192.168.0.1")
{:ok,
%IPA.Address{address: "192.168.0.1",
binary: "0b11000000101010000000000000000001",
bits: "11000000.10101000.00000000.00000001",
block: :rfc1918,
hex: "0xC0A80001",
octets: {192, 168, 0, 1},
reserved: true,
version: 4}}
iex> ip.address
"192.168.0.1"
iex> ip.binary
"0b11000000101010000000000000000001"
iex> ip.bits
"11000000.10101000.00000000.00000001"
iex> ip.block
:rfc1918
iex> ip.hex
"0xC0A80001"
iex> ip.octets
{192, 168, 0, 1}
iex> ip.version
4
iex> ip.reserved
true
iex> IPA.address("192.168.0.256")
{:error, :invalid_ipv4_address}
iex> IPA.address("192.168.0")
{:error, :invalid_ipv4_address}

address!(addr)

Specs

address!(addr) :: addr_struct | no_return

Same as address/1 but raises IPError if it fails, returns the IP Address struct otherwise.

block(addr)

Specs

block(addr) :: atom

Returns an atom describing which reserved block the address is a member of if it is a private address, returns :public otherwise.

Available blocks:

:this_network - 0.0.0.0/8 Used for broadcast messages to the current “this” network as specified by RFC 1700, page 4.

:rfc1918 - 10.0.0.0/8, 172.16.0.0/12 & 192.168.0.0/16 Used for local communications within a private network as specified by RFC 1918.

:rfc6598 - 100.64.0.0/10 Used for communications between a service provider and its subscribers when using a Carrier-grade NAT, as specified by RFC 6598.

:loopback - 127.0.0.0/8 Used for loopback addresses to the local host, as specified by RFC 990.

:link_local - 169.254.0.0/16 Used for link-local addresses between two hosts on a single link when no IP address is otherwise specified, such as would have normally been retrieved from a DHCP server, as specified by RFC 3927.

:rfc5736 - 192.0.0.0/24 Used for the IANA IPv4 Special Purpose Address Registry as specified by RFC 5736.

:rfc5737 - 192.0.2.0/24, 198.51.100.0/24 & 203.0.113.0/24 Assigned as “TEST-NET” in RFC 5737 for use solely in documentation and example source code and should not be used publicly.

:rfc3068 - 192.88.99.0/24 Used by 6to4 anycast relays as specified by RFC 3068.

:rfc2544 - 198.18.0.0/15 Used for testing of inter-network communications between two separate subnets as specified in RFC 2544.

:multicast - 224.0.0.0/4 Reserved for multicast assignments as specified in RFC 5771. 233.252.0.0/24 is assigned as “MCAST-TEST-NET” for use solely in documentation and example source code.

:future - 240.0.0.0/4 Reserved for future use, as specified by RFC 6890.

:limited_broadcast - 255.255.255.255/32 Reserved for the “limited broadcast” destination address, as specified by RFC 6890.

:public - All other addresses are public.

Examples

iex> IPA.Address.block(“8.8.8.8”) :public iex> IPA.Address.block(“192.168.0.1”) :rfc1918

reserved?(addr)

Specs

reserved?(addr) :: boolean

Checks whether a given IP address is reserved.

to_binary(addr)

Specs

to_binary(addr) :: String.t

Converts a dot decimal IP address to a 0b prefixed binary number.

Example

iex> IPA.Address.to_binary("192.168.0.1")
"0b11000000101010000000000000000001"

to_bits(addr)

Specs

to_bits(addr) :: String.t

Converts a dot decimal IP address to binary bits.

Example

iex> IPA.Address.to_bits("192.168.0.1")
"11000000.10101000.00000000.00000001"

to_hex(addr)

Specs

to_hex(addr) :: String.t

Converts a dot decimal IP address to a 0x prefixed hexadecimal number.

Example

iex> IPA.Address.to_hex("192.168.0.1")
"0xC0A80001"

to_octets(addr)

Specs

to_octets(addr) :: {integer}

Converst a dot decimal IP address to a 4 element tuple, representing the given addresses 4 octets.

Example

iex> IPA.Address.to_octets("192.168.0.1")
{192, 168, 0, 1}

valid?(addr)

Specs

valid?(addr) :: boolean

Checks if the given address is a valid IP address.

Uses a regular expression to check there is exactly 4 integers between 0 & 255, inclusive, separated by dots. Taken from here. Therefore does not currently take into consideration the fact that 127.1 can be considered a valid IP address that translates to 127.0.0.1.